+++ src/lemlib/chassis/pursuit.cpp
@@ -145,7 +145,7 @@ lemlib::Pose lookaheadPoint(lemlib::Pose lastLookahead, lemlib::Pose pose, std::
                             float lookaheadDist) {
     // initialize variables
     lemlib::Pose lookahead = lastLookahead;
-    double t;
+    float t;
 
     // find the furthest lookahead point on the path
     for (int i = 0; i < path.size() - 1; i++) {
@@ -165,16 +165,16 @@ lemlib::Pose lookaheadPoint(lemlib::Pose lastLookahead, lemlib::Pose pose, std::
  * @param pos the position of the robot
  * @param heading the heading of the robot
  * @param lookahead the lookahead point
- * @return double curvature
+ * @return float curvature
  */
-double findLookaheadCurvature(lemlib::Pose pose, double heading, lemlib::Pose lookahead) {
+float findLookaheadCurvature(lemlib::Pose pose, float heading, lemlib::Pose lookahead) {
     // calculate whether the robot is on the left or right side of the circle
-    double side = lemlib::sgn(std::sin(heading) * (lookahead.x - pose.x) - std::cos(heading) * (lookahead.y - pose.y));
+    float side = lemlib::sgn(std::sin(heading) * (lookahead.x - pose.x) - std::cos(heading) * (lookahead.y - pose.y));
     // calculate center point and radius
-    double a = -std::tan(heading);
-    double c = std::tan(heading) * pose.x - pose.y;
-    double x = std::fabs(a * lookahead.x + lookahead.y + c) / std::sqrt((a * a) + 1);
-    double d = std::hypot(lookahead.x - pose.x, lookahead.y - pose.y);
+    float a = -std::tan(heading);
+    float c = std::tan(heading) * pose.x - pose.y;
+    float x = std::fabs(a * lookahead.x + lookahead.y + c) / std::sqrt((a * a) + 1);
+    float d = std::hypot(lookahead.x - pose.x, lookahead.y - pose.y);
 
     // return curvature
     return side * ((2 * x) / (d * d));
@@ -197,7 +197,7 @@ void lemlib::Chassis::follow(asset path, int timeout, float lookahead, bool reve
     Pose lookaheadPose(0, 0, 0);
     Pose lastLookahead = pathPoints.at(0);
     lastLookahead.theta = 0;
-    double curvature;
+    float curvature;
     float targetVel;
     float prevLeftVel = 0;
     float prevRightVel = 0;
@@ -222,7 +222,7 @@ void lemlib::Chassis::follow(asset path, int timeout, float lookahead, bool reve
         lastLookahead = lookaheadPose; // update last lookahead position
 
         // get the curvature of the arc between the robot and the lookahead point
-        double curvatureHeading = M_PI / 2 - pose.theta;
+        float curvatureHeading = M_PI / 2 - pose.theta;
         curvature = findLookaheadCurvature(pose, curvatureHeading, lookaheadPose);
 
         // get the target velocity of the robot
